C vector length and indices (bootstrap)

+++ bde/bmem/lib/alloc.c
@@ -3,8 +3,8 @@
 /*    -------------------------------------------------------------    */
 /*    Author      :  Manuel Serrano                                    */
 /*    Creation    :  Sun Apr 13 06:42:57 2003                          */
-/*    Last change :  Wed Dec  7 13:58:01 2016 (serrano)                */
-/*    Copyright   :  2003-16 Manuel Serrano                            */
+/*    Last change :  Thu Jul 27 08:16:16 2017 (serrano)                */
+/*    Copyright   :  2003-17 Manuel Serrano                            */
 /*    -------------------------------------------------------------    */
 /*    Allocation replacement routines                                  */
 /*=====================================================================*/
@@ -769,13 +769,13 @@ WRAPPER( string_to_keyword, KEYWORD_TYPE_NUM, (char *s), (s) )
 WRAPPER( bstring_to_keyword, KEYWORD_TYPE_NUM, (void *s), (s) )
 
 /* vector */
-WRAPPER( create_vector, VECTOR_TYPE_NUM, (int len), (len) )
-WRAPPER( make_vector, VECTOR_TYPE_NUM, (int len, obj_t init), (len, init) )
-WRAPPER( make_vector_uncollectable, VECTOR_TYPE_NUM, (int len, obj_t init), (len, init) )
+WRAPPER( create_vector, VECTOR_TYPE_NUM, (long len), (len) )
+WRAPPER( make_vector, VECTOR_TYPE_NUM, (long len, obj_t init), (len, init) )
+WRAPPER( make_vector_uncollectable, VECTOR_TYPE_NUM, (long len, obj_t init), (len, init) )
 
 /* create_vector_uncollectable is only */ 
 /* used to allocate class objects.     */
-WRAPPER( create_vector_uncollectable, CLASS_TYPE_NUM, (int len), (len) )
+WRAPPER( create_vector_uncollectable, CLASS_TYPE_NUM, (long len), (len) )
 
 /* procedure */
 WRAPPER( make_fx_procedure, PROCEDURE_TYPE_NUM, (obj_t (*e)(), int a, int s), ((void *(*)())e, a, s) )
